Build Mapping Table

Builds a mapping table between gene symbol and node names in the HPRD interactome network.

buildGeneIntTable(intome)

Arguments

intome
interactome (HPRD (Human Protein Reference Database) derived from the DLBCL package.)

Details

This function takes the HPRD protein protein network that is available in the DLBCL package and produces a mapping table that is needed by functions in the surrogateMutation package.

Value

A mapping table with the following columns:

NodeName

Node Names in the interactome network. These node names are in the format of GeneSymbol(EntrezID)

Gene

Gene Symbols, extracted from the NodeName.

EntrezID

Entrez Gene IDs, extracted from the NodeName.

References

Note

See also

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
library(DLBCL)
data(interactome)
geneIntTable <- buildGeneIntTable(interactome)

#> Warning: some row.names duplicated: 8699,9178 --> row.names NOT used

geneIntTable[1:5,]

#>       NodeName     Gene EntrezID
#> 1 SERPINA3(12) SERPINA3       12
#> 2     ABAT(18)     ABAT       18
#> 3     ABL1(25)     ABL1       25
#> 4    ACADM(34)    ACADM       34
#> 5   ACADVL(37)   ACADVL       37