plotAllGraphs

Given a surrogateTable object generated by buildSurrogateTable(), generate all surrogate graphs.

plotAllGraphs(surrogateTable, gisticCopyCalls, surrogateResults, intome)

Arguments

surrogateTable
A surrogateTable data frame generated from buildSurrogateTable()

gisticCopyCalls
When set to NULL, plotAllGraphs will look in the surrogateResults$mutCopyFrames object for the copy number calls for each gene.

surrogateResults

intome

Details

Value

An SVG file generated for each line in the surrogateTable data frame. The SVG files will have the format (Sample)-GeneName.svg.

References

Note

See also

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
##load surrogate result object
library(DLBCL)
data(interactome)

data(surrogateResult)
surrogateTable <- buildSurrogateTable(surrogateResult)

##filter the surrogate table based on p-value and return surrogate mutations greater than 5.
filteredSurrTable <- surrogateTable[surrogateTable$pvalue < 0.05 & surrogateTable$neighbor > 3,]
filteredSurrTable

#>                                        ID    Gene       Sample      NodeName
#> TCGA.AO.A12D-IRS2       TCGA.AO.A12D-IRS2    IRS2 TCGA.AO.A12D    IRS2(8660)
#> TCGA.AO.A12D-PIK3R3   TCGA.AO.A12D-PIK3R3  PIK3R3 TCGA.AO.A12D  PIK3R3(8503)
#> TCGA.AO.A12D-RAF1       TCGA.AO.A12D-RAF1    RAF1 TCGA.AO.A12D    RAF1(5894)
#> TCGA.AO.A12D-SKP2       TCGA.AO.A12D-SKP2    SKP2 TCGA.AO.A12D    SKP2(6502)
#> TCGA.AO.A12D-SMARCA4 TCGA.AO.A12D-SMARCA4 SMARCA4 TCGA.AO.A12D SMARCA4(6597)
#> TCGA.AO.A12D-YWHAE     TCGA.AO.A12D-YWHAE   YWHAE TCGA.AO.A12D   YWHAE(7531)
#> TCGA.BH.A0DS-CSNK2A1 TCGA.BH.A0DS-CSNK2A1 CSNK2A1 TCGA.BH.A0DS CSNK2A1(1457)
#> TCGA.BH.A0DS-GSK3B     TCGA.BH.A0DS-GSK3B   GSK3B TCGA.BH.A0DS   GSK3B(2932)
#> TCGA.BH.A0DS-GTF2H1   TCGA.BH.A0DS-GTF2H1  GTF2H1 TCGA.BH.A0DS  GTF2H1(2965)
#> TCGA.BH.A0DS-IRS4       TCGA.BH.A0DS-IRS4    IRS4 TCGA.BH.A0DS    IRS4(8471)
#> TCGA.BH.A0DS-MAPK14   TCGA.BH.A0DS-MAPK14  MAPK14 TCGA.BH.A0DS  MAPK14(1432)
#> TCGA.E2.A1B5-GSK3B     TCGA.E2.A1B5-GSK3B   GSK3B TCGA.E2.A1B5   GSK3B(2932)
#> TCGA.E2.A1B5-GTF2H1   TCGA.E2.A1B5-GTF2H1  GTF2H1 TCGA.E2.A1B5  GTF2H1(2965)
#> TCGA.E2.A1B5-HDAC3     TCGA.E2.A1B5-HDAC3   HDAC3 TCGA.E2.A1B5   HDAC3(8841)
#> TCGA.E2.A1B5-HSPCA     TCGA.E2.A1B5-HSPCA   HSPCA TCGA.E2.A1B5   HSPCA(3320)
#> TCGA.E2.A1B5-MAPK8     TCGA.E2.A1B5-MAPK8   MAPK8 TCGA.E2.A1B5   MAPK8(5599)
#> TCGA.E2.A1B5-PDGFRB   TCGA.E2.A1B5-PDGFRB  PDGFRB TCGA.E2.A1B5  PDGFRB(5159)
#> TCGA.E2.A1B5-PTK2       TCGA.E2.A1B5-PTK2    PTK2 TCGA.E2.A1B5    PTK2(5747)
#> TCGA.E2.A1B5-UBB         TCGA.E2.A1B5-UBB     UBB TCGA.E2.A1B5     UBB(7314)
#>                      neighbor degree pvalue isMutated
#> TCGA.AO.A12D-IRS2           4     35  0.025         1
#> TCGA.AO.A12D-PIK3R3         4     11  0.000         0
#> TCGA.AO.A12D-RAF1           7     89  0.023         0
#> TCGA.AO.A12D-SKP2           5     25  0.002         0
#> TCGA.AO.A12D-SMARCA4        5     36  0.004         0
#> TCGA.AO.A12D-YWHAE          7     58  0.001         0
#> TCGA.BH.A0DS-CSNK2A1       14    159  0.009         0
#> TCGA.BH.A0DS-GSK3B          6     59  0.028         0
#> TCGA.BH.A0DS-GTF2H1         4     26  0.014         0
#> TCGA.BH.A0DS-IRS4           4     14  0.002         0
#> TCGA.BH.A0DS-MAPK14         7     76  0.043         0
#> TCGA.E2.A1B5-GSK3B          5     59  0.035         0
#> TCGA.E2.A1B5-GTF2H1         4     26  0.007         0
#> TCGA.E2.A1B5-HDAC3          5     57  0.028         0
#> TCGA.E2.A1B5-HSPCA          6     74  0.019         0
#> TCGA.E2.A1B5-MAPK8          6     63  0.005         0
#> TCGA.E2.A1B5-PDGFRB         4     53  0.046         0
#> TCGA.E2.A1B5-PTK2           6     81  0.020         0
#> TCGA.E2.A1B5-UBB            5     55  0.022         0

#plot all graphs in current working directory
#be careful with the size of the table you pass to plotAllGraphs - it may generate
#a huge number of files.
plotAllGraphs(filteredSurrTable, surrogateResult, intome=interactome)

#> [1] "IRS2(8660) TCGA.AO.A12D"

#> Error in renderSubNetSimple(NodeName, Sample, GeneName, intome, gisticCopyCalls = NULL,     resultObj = surrogateResults): object 'geneIntTable' not found